requirements.txt
psutil
discord.py